Refactor Game 2 into different includes
