+++ Prototyping/Games/2 - Circle shoots Cubes/Game 2 - Circle shoots Cubes.cpp
@@ -1,385 +1,14 @@
 #include <QApplication>
-#include <QColor>
-#include <QDebug>
-#include <QGraphicsEllipseItem>
-#include <QGraphicsObject>
-#include <QGraphicsRectItem>
-#include <QGraphicsScene>
-#include <QGraphicsSceneMouseEvent>
-#include <QGraphicsView>
-#include <QKeyEvent>
-#include <QPainter>
-#include <QPropertyAnimation>
-#include <QTimer>
-#include <functional>
-#include <map>
-#include <random>
 
-struct RGBColor {
-    int red, green, blue;
-};
-
-class ClickableRectItem : public QGraphicsRectItem {
-    std::function<void()> _onClickHandler      = []() {};
-    std::function<void()> _onRightClickHandler = []() {};
-
-public:
-    ClickableRectItem(const QRectF& rect, QGraphicsItem* parent = nullptr)
-        : QGraphicsRectItem(rect, parent) {}
-
-    void OnClick(std::function<void()> handler) { _onClickHandler = handler; }
-    void OnRightClick(std::function<void()> handler) { _onRightClickHandler = handler; }
-
-protected:
-    void mousePressEvent(QGraphicsSceneMouseEvent* event) override {
-        if (event->button() == Qt::LeftButton) {
-            _onClickHandler();
-        } else if (event->button() == Qt::RightButton) {
-            _onRightClickHandler();
-        }
-        QGraphicsRectItem::mousePressEvent(event);
-    }
-};
-
-class AnimatedEllipse : public QGraphicsObject {
-public:
-    AnimatedEllipse(qreal x, qreal y, qreal w, qreal h) : rect(x, y, w, h) { setPos(x, y); }
-
-    QRectF boundingRect() const override { return rect; }
-
-    void paint(QPainter* painter, const QStyleOptionGraphicsItem* option, QWidget* widget)
-        override {
-        painter->setBrush(Qt::blue);
-        painter->drawEllipse(rect);
-    }
-
-private:
-    QRectF rect;
-};
-
-class GridMap {
-    std::pair<int, int>              _currentCirclePosition;
-    std::vector<std::pair<int, int>> _currentCubePositions;
-
-public:
-    GridMap(int rows, int cols) : rows(rows), cols(cols), cellWidth(50), cellHeight(50) {
-        scene      = new QGraphicsScene();
-        tileColors = std::map<std::pair<int, int>, QColor>();
-        // drawGrid();
-    }
-
-    QGraphicsScene* getScene() const { return scene; }
-
-    void SetTileLines(bool state) { drawTileLines = state; }
-
-    void SetTileColor(int row, int col, RGBColor color) {
-        QColor qcolor(color.red, color.green, color.blue);
-        tileColors[{row, col}] = qcolor;
-    }
-
-    void AddCube(int row, int col) { _currentCubePositions.push_back({row, col}); }
-
-    void MoveCircleTo(int row, int col) { _currentCirclePosition = {row, col}; }
-
-    void MoveCircleUp() {
-        if (_currentCirclePosition.first > 0)
-            MoveCircleTo(_currentCirclePosition.first - 1, _currentCirclePosition.second);
-    }
-    void MoveCircleDown() {
-        if (_currentCirclePosition.first < rows - 1)
-            MoveCircleTo(_currentCirclePosition.first + 1, _currentCirclePosition.second);
-    }
-    void MoveCircleLeft() {
-        if (_currentCirclePosition.second > 0)
-            MoveCircleTo(_currentCirclePosition.first, _currentCirclePosition.second - 1);
-    }
-    void MoveCircleRight() {
-        if (_currentCirclePosition.second < cols - 1)
-            MoveCircleTo(_currentCirclePosition.first, _currentCirclePosition.second + 1);
-    }
-
-    void Redraw() {
-        scene->clear();
-        drawGrid();
-        drawCubes();
-        drawCircleOnTile(
-            _currentCirclePosition.first, _currentCirclePosition.second, {207, 2, 152}
-        );
-    }
-
-private:
-    std::pair<int, int> gridCoordinates(QPointF scenePos) {
-        int gridX = static_cast<int>(scenePos.x()) / cellWidth;
-        int gridY = static_cast<int>(scenePos.y()) / cellHeight;
-        return {gridY, gridX};
-    }
-
-    std::map<std::pair<int, int>, QGraphicsRectItem*>& getCubeItems() { return _currentCubeItems; }
-
-    void explodeUnderProjectileIfAnyCubes(QPointF pos) {
-        auto foundCube = getCubeItems().find(gridCoordinates(pos));
-        if (foundCube != _currentCubeItems.end()) {
-            qDebug() << "Found cube to explode";
-            // explode in a single shot
-            // QTimer::singleShot(0, [this, foundCube]() { explode(foundCube->second); });
-            explode(foundCube->second);
-            _currentCubeItems.erase(foundCube);
-            _currentCubePositions.erase(
-                std::remove(
-                    _currentCubePositions.begin(), _currentCubePositions.end(), gridCoordinates(pos)
-                ),
-                _currentCubePositions.end()
-            );
-        }
-    }
-
-    void fireProjectile(std::pair<int, int> from, std::pair<int, int> to) {
-        qDebug() << "Firing projectile from" << from.first << "," << from.second << "to" << to.first
-                 << "," << to.second;
-
-        auto startingPoint = std::make_pair(
-            from.second * cellHeight + cellHeight / 2, from.first * cellWidth + cellWidth / 2
-        );
-        auto endingPoint = std::make_pair(
-            to.first * cellHeight + cellHeight / 2, to.second * cellWidth + cellWidth / 2
-        );
-
-        // Create a small ellipse representing the projectile
-        AnimatedEllipse* projectile = new AnimatedEllipse(0, 0, 10, 10);
-
-        QTimer* timer = new QTimer();
-
-        QObject::connect(timer, &QTimer::timeout, [projectile, this]() {
-            QPointF scenePos = projectile->pos();
-            explodeUnderProjectileIfAnyCubes(scenePos);
-        });
-
-        // Add the projectile to the scene
-        scene->addItem(projectile);
-
-        QObject* projectilePtr = static_cast<QObject*>(projectile);
-
-        // Create a property animation for the x and y coordinates
-        QPropertyAnimation* animation = new QPropertyAnimation(projectile, "pos");
-        animation->setDuration(1000);  // Animation duration in milliseconds
-        animation->setStartValue(QPointF(startingPoint.first, startingPoint.second));
-        animation->setEndValue(QPointF(endingPoint.second, endingPoint.first));
-
-        // Connect the finished signal of the animations to a lambda function that removes and
-        // deletes the projectile
-        QObject::connect(animation, &QPropertyAnimation::finished, [projectile, this, to]() {
-            explodeUnderProjectileIfAnyCubes(projectile->pos());
-            scene->removeItem(projectile);
-            delete projectile;
-            qDebug() << "Projectile reached" << to.first << "," << to.second;
-        });
-
-        // Start the animations
-        animation->start(QAbstractAnimation::DeleteWhenStopped);
-        timer->start(50);
-
-        QObject::connect(animation, &QPropertyAnimation::finished, [timer]() {
-            timer->stop();
-            delete timer;
-        });
-    }
-
-    std::pair<int, int> getCurrentCirclePosition() { return _currentCirclePosition; }
-
-    void explode(QGraphicsRectItem* cube) {
-        if (!cube) {
-            qDebug() << "Cube is null";
-            return;
-        }
-        if (cube->x() == 0) {
-            qDebug() << "Something is wrong, the cube is at 0,0";
-        }
-
-        std::default_random_engine            generator;
-        std::uniform_real_distribution<qreal> distribution(-50.0, 50.0);
-
-        qDebug() << "Explode cube which is at " << cube->x() << "," << cube->y();
-
-        // Create a bunch of particles at the cube's location
-        const int numParticles = 100;
-        for (int i = 0; i < numParticles; ++i) {
-            auto particle = new AnimatedEllipse(0, 0, 5, 5);
-            particle->setPos(cube->scenePos());
-            scene->addItem(particle);
-
-            // Create a property animation for the x and y coordinates
-            QPropertyAnimation* animation = new QPropertyAnimation(particle, "pos");
-            animation->setDuration(500);  // Animation duration in milliseconds
-
-            // Set the end position to a random offset from the cube's position
-            qreal endX = cube->x() + distribution(generator);
-            qreal endY = cube->y() + distribution(generator);
-            animation->setEndValue(QPointF(endX, endY));
-
-            // Connect the finished signal of the animations to a lambda function that removes and
-            // deletes the particle
-            QObject::connect(animation, &QPropertyAnimation::finished, [particle, this]() {
-                scene->removeItem(particle);
-                delete particle;
-            });
-
-            // Start the animation
-            animation->start(QAbstractAnimation::DeleteWhenStopped);
-        }
-
-        // Hide or delete the cube
-        scene->removeItem(cube);
-        delete cube;
-    }
-
-    void drawGrid() {
-        for (int i = 0; i < rows; ++i) {
-            for (int j = 0; j < cols; ++j) {
-                int offsetX = j * cellWidth;
-                int offsetY = i * cellHeight;
-
-                QRectF rect(offsetX, offsetY, cellWidth, cellHeight);
-
-                ClickableRectItem* item = new ClickableRectItem(rect);
-                item->OnClick([this, i, j]() {
-                    qDebug() << "Clicked on tile" << i << j;
-                    qDebug() << "Current circle position is" << getCurrentCirclePosition();
-                    auto from = getCurrentCirclePosition();
-                    fireProjectile(from, {i, j});
-                });
-                item->OnRightClick([this, i, j]() {
-                    // Single shot timer
-                    QTimer::singleShot(0, [this, i, j]() { AddCube(i, j); });
-                });
-                if (drawTileLines) {
-                    item->setPen(QPen(Qt::black, 1));
-                }
-                if (tileColors.find({i, j}) != tileColors.end()) {
-                    item->setBrush(tileColors[{i, j}]);
-                } else {
-                    item->setBrush(Qt::white);
-                }
-
-                scene->addItem(item);
-            }
-        }
-    }
-
-    std::map<std::pair<int, int>, QGraphicsRectItem*> _currentCubeItems;
-
-    void drawCubes() {
-        _currentCubeItems.clear();
-        for (auto& cube : _currentCubePositions) {
-            int offsetX = cube.second * cellWidth;
-            int offsetY = cube.first * cellHeight;
-
-            int                size = cellWidth / 2;  // You can adjust these values as you need
-            QGraphicsRectItem* item = new QGraphicsRectItem(0, 0, size, size);
-            item->setPos(offsetX + size / 2, offsetY + size / 2);
-
-            qDebug() << "Adding current cube item at " << cube.first << "," << cube.second;
-            _currentCubeItems[{cube.first, cube.second}] = item;
-            item->setBrush(Qt::blue);
-            scene->addItem(item);
-        }
-    }
-
-    void drawCircleOnTile(int row, int col, RGBColor color) {
-        QColor qcolor(color.red, color.green, color.blue);
-        int    offsetX = col * cellWidth;
-        int    offsetY = row * cellHeight;
-
-        int circleWidth  = cellWidth / 2;  // You can adjust these values as you need
-        int circleHeight = cellHeight / 2;
-
-        // Position circle in the middle of the cell
-        offsetX += cellWidth / 4;
-        offsetY += cellHeight / 4;
-
-        QGraphicsEllipseItem* item =
-            new QGraphicsEllipseItem(offsetX, offsetY, circleWidth, circleHeight);
-        item->setBrush(qcolor);
-
-        scene->addItem(item);
-    }
-
-    int                                   rows;
-    int                                   cols;
-    int                                   cellWidth;
-    int                                   cellHeight;
-    QGraphicsScene*                       scene;
-    bool                                  drawTileLines;
-    std::map<std::pair<int, int>, QColor> tileColors;
-};
-
-class KeyPressGraphicsView : public QGraphicsView {
-    // GridMap* _gridMap;
-
-public:
-    KeyPressGraphicsView(QGraphicsScene* scene, QWidget* parent = nullptr)
-        : QGraphicsView(scene, parent) {}
-
-    // : _gridMap(map), QGraphicsView(scene, parent) {}
-
-    // protected:
-    //     void OnUpKey() {
-    //         if (_gridMap) _gridMap->MoveCircleUp();
-    //     }
-    //     void OnDownKey() {
-    //         if (_gridMap) _gridMap->MoveCircleDown();
-    //     }
-    //     void OnLeftKey() {
-    //         if (_gridMap) _gridMap->MoveCircleLeft();
-    //     }
-    //     void OnRightKey() {
-    //         if (_gridMap) _gridMap->MoveCircleRight();
-    //     }
-
-    // void keyPressEvent(QKeyEvent* event) override {
-    //     switch (event->key()) {
-    //         case Qt::Key_Up:
-    //             OnUpKey();
-    //             break;
-    //         case Qt::Key_Down:
-    //         case Qt::Key_S:
-    //             OnDownKey();
-    //             break;
-    //         case Qt::Key_Left:
-    //         case Qt::Key_A:
-    //             OnLeftKey();
-    //             break;
-    //         case Qt::Key_Right:
-    //         case Qt::Key_D:
-    //             OnRightKey();
-    //             break;
-    //         default:
-    //             QGraphicsView::keyPressEvent(event);
-    //             break;
-    //     }
-    // }
-};
+#include "Prototyping/GameGraphicsView.h"
+#include "Prototyping/GameMapForCircleAndCubes.h"
 
 int main(int argc, char* argv[]) {
-    int z = 420;
-    int x = 69;
-
-    QApplication app{argc, argv};
-
-    GridMap map(10, 15);
-    // map.SetTileLines(true);
-    // map.MoveCircleTo(5, 5);
-    // map.AddCube(1, 2);
-    // map.AddCube(8, 7);
-    // map.AddCube(3, 8);
-
-    auto scene = map.getScene();
-
-    KeyPressGraphicsView view(scene);
+    QApplication                          app{argc, argv};
+    Prototyping::GameMapForCircleAndCubes map{10, 20};
+    Prototyping::GameGraphicsView         view;
+    view.SetScene(map.GetScene());
+    map.Redraw();
     view.show();
-
-    // map.Redraw();
-
-    app.exec();
-
-    return 0;
+    return app.exec();
 }
