+++ Prototyping/Games/2 - Circle shoots Cubes/include/Prototyping/RBGColor.h
@@ -1,0 +1,8 @@
+#pragma once
+
+namespace Prototyping {
+
+    struct RGBColor {
+        int red, green, blue;
+    };
+}
