+++ Prototyping/Games/2 - Circle shoots Cubes/xmake.lua
@@ -4,3 +4,4 @@ target("Game 2 - Circle shoots Cubes")
     add_qt()
     add_string_formatting()
     add_logging_support_for_binaries()
+    add_includedirs("include")
